requirements.txt
odfpy==1.4.1
python-docx==0.8.11
tk==0.1.0
tkscrolledframe==1.0.4